Large refactor
