+++ moosebot/__init__.py
@@ -1,0 +1,12 @@
+from moosebot.MooseBot import MooseBot
+from moosebot.database import MooseDb
+
+__all__ = [
+    "cogs",
+    "commands",
+    "converters",
+    "MooseBot.py",
+    "MooseBot",
+    "MooseDb",
+    "utils"
+]
