+++ moosebot/cogs/__init__.py
@@ -1,0 +1,7 @@
+from moosebot.cogs.Info import Info
+from moosebot.cogs.Counting import Counting
+
+__all__ = [
+    "Info",
+    "Counting"
+]
