+++ moosebot/converters/__init__.py
@@ -1,0 +1,14 @@
+from moosebot.converters.VoiceChannel import VoiceChannel
+from moosebot.converters.FullMember import FullMember
+from moosebot.converters.PartialMember import PartialMember
+from moosebot.converters.Role import Role
+from moosebot.converters.Url import Url
+
+
+__all__ = [
+    "FullMember",
+    "PartialMember",
+    "Role",
+    "Url",
+    "VoiceChannel"
+]
