+++ moosebot/database.py
@@ -1,0 +1,9 @@
+from pymongo import MongoClient
+
+
+class MooseDb:
+
+    def __init__(self):
+        self.client = MongoClient()
+        self.db = self.client.MooseBot
+        # TODO finish
